pyproject.toml
[project]
name = "aitudes"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-adk>=1.0.0",
    "huggingface-hub[mcp]>=0.32.0",
    "jinja2>=3.1.6",
    "loguru>=0.7.3",
    "numpy>=2.2.5",
    "onnxruntime>=1.21.1",
    "openai>=1.75.0",
    "openai-agents>=0.0.12",
    "pandas>=2.2.3",
    "polars>=1.27.1",
    "python-dotenv>=1.1.0",
    "rich>=14.0.0",
    "smolagents[toolkit]>=1.16.1",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "nbstripout>=0.8.1",
    "pre-commit>=4.2.0",
    "ruff>=0.11.6",
]
